Delete replies system if post/user is deleted

+++ groupomania-back-end/controllers/user.js
@@ -1,6 +1,7 @@
 const db = require("../models/index.js");
 const Users = db.users; // import 'users' table
 const Posts = db.posts; // import 'posts' table
+const Reply = db.reply; // import 'reply' table
 
 const fs = require('fs');
 const jwt = require('jsonwebtoken');
@@ -104,22 +105,40 @@ exports.setUserData = (req, res, next) => { // Set data of user
 
 exports.deleteUser = (req, res, next) => { // Delete account
     const isDeleted = 1;
-    Posts.destroy({ where: { ownerId: getUserIdFromRequest(req) }, force: true })
-        .then(result => { // Delete all posts of user
-            console.log(result + ' posts a/ont été delete lors de la suppresion d\'un compte.')
-            Users.findOne({ where: {id: getUserIdFromRequest(req)} })
-                .then(data => {
-                    if(data.imgProfil) { // Delete avatar file (in storage)
-                        const lastUserImg = data.imgProfil.split('/images/')[1];
-                        fs.unlink('images/' + lastUserImg, () => {});
-                    }
-                    Users.destroy({ where: { id: getUserIdFromRequest(req) }, force: true })
-                        .then(result => { // Delete accout (final step)
-                            if(result == isDeleted) {
-                                res.status(201).send({message: 'Votre compte a correctement été supprimé.'})
-                            } else { throw 'Votre compte n\'a pas été supprimé.' }
-                        })
-                        .catch(err => { res.status(409).send({message: 'Une erreur est survenue (' + err + ')'}) });
+    Users.findOne({where: {id: getUserIdFromRequest(req)}})
+        .then(data => {
+            if(data.imgProfil) { // Delete avatar file (in storage)
+                const lastUserImg = data.imgProfil.split('/images/')[1];
+                fs.unlink('images/' + lastUserImg, () => {});
+            }
+            function deleteRepliesAndImg() {
+                return new Promise(resolve => {
+                    Posts.findAll({where: {ownerId: getUserIdFromRequest(req)}})
+                    .then(postData => { // Delete all replies & images from user posts
+                        for(let i = 0; i < postData.length; i++) {
+                            Reply.destroy({where: {postId: postData[i].dataValues.id}, force: true}) 
+                            if(postData[i].dataValues.fileImg) {
+                                const postImg = postData[i].dataValues.fileImg.split('/images/')[1];
+                                fs.unlink('images/' + postImg, () => {});
+                                if(i == postData.length-1) { resolve(true) }
+                            }
+                        }
+                    });
+                });
+            }
+            async function deleteRessources() {
+                var isResolve = await deleteRepliesAndImg();
+                if(isResolve == true) { // Delete all posts and all replies from user
+                    Posts.destroy({where: {ownerId: getUserIdFromRequest(req)}, force: true})
+                    Reply.destroy({where: {ownerId: getUserIdFromRequest(req)}, force: true})
+                }
+            }
+            deleteRessources();
+            Users.destroy({where: { id: getUserIdFromRequest(req) }, force: true})
+                .then(result => { // Delete accout (final step)
+                    if(result == isDeleted) {
+                        res.status(201).send({message: 'Votre compte a correctement été supprimé.'})
+                    } else { throw 'Votre compte n\'a pas été supprimé.' }
                 })
                 .catch(err => { res.status(409).send({message: 'Une erreur est survenue (' + err + ')'}) });
         })
